# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScopeCreate": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "kind": {
            "default": "standard",
            "enum": [
              "standard",
              "drafting"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label"
        ],
        "title": "ScopeCreate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Authenticated C1 HTTP API. Every route except GET /v1/readyz requires an OIDC bearer access token from the configured issuer. Request bodies are limited to 1 MiB; unknown query parameters, duplicate parameters and backend selectors are rejected with 400. Errors are RFC 9457 Problem Details. Generated by scripts/openapi.py; not served by C1.",
    "title": "C1 API",
    "version": "0.1.0rc2"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/access-scopes": {
      "get": {
        "operationId": "list_scopes_v1_access_scopes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "title": "Response List Scopes V1 Access Scopes Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List Scopes"
      },
      "post": {
        "operationId": "create_scope_v1_access_scopes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Scope V1 Access Scopes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Scope"
      }
    },
...
    "/v1/assertions": {
      "get": {
        "operationId": "query_assertions_v1_assertions_get",
        "parameters": [
          {
            "in": "query",
            "name": "competing_for",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_unknown",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "object",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "predicate",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "review_state",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "valid_at",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Query Assertions V1 Assertions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Query Assertions"
      }
    },
    "/v1/catalog": {
      "get": {
        "operationId": "query_catalog_v1_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Query Catalog V1 Catalog Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Query Catalog"
      }
    },
...
    "/v1/context": {
      "post": {
        "operationId": "context_build_v1_context_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Context Build V1 Context Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Context Build"
      }
    },
    "/v1/documents": {
      "get": {
        "operationId": "documents_list_v1_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "text_contains",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Documents List V1 Documents Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Documents List"
      }
    },
    "/v1/documents/by-id": {
      "get": {
        "description": "Read an IRI without interpreting its final segment as an operation.",
        "operationId": "document_read_by_id_v1_documents_by_id_get",
        "parameters": [
          {
            "in": "query",
            "name": "document_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Document Read By Id V1 Documents By Id Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Document Read By Id"
      }
    },
...
    "/v1/entities/search": {
      "post": {
        "operationId": "query_entities_search_v1_entities_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryFilters"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Query Entities Search V1 Entities Search Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Query Entities Search"
      }
    },
...
    "/v1/setup/confirm": {
      "post": {
        "operationId": "setup_confirm_v1_setup_confirm_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Not found, or not available to the caller"
          },
          "503": {
            "description": "Not ready, or a required service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Setup Confirm"
      }
    },
...
  }
}